Moved team invitations model to team collaborator model
