+++ src/app/api/my/team/invitation/route.ts
@@ -1,32 +1,9 @@
 import _ from 'lodash';
 
-import {
-  getTeamInvitations,
-  invitePersonToMyTeam,
-} from '@/controllers/teamInvitation.controller';
+import { invitePersonToMyTeam } from '@/controllers/teamCollaborator.controller';
 import { AuthenticationMiddleware } from '@/middlewares/authentication.middleware';
-import { getPaginationFromParams } from '@/utils/paginateData';
 import { isErrorWithMessage } from '@/utils/error.utils';
 import { User } from '@/models/user.model';
-
-export const GET = async (request: NextRequest) => {
-  try {
-    await AuthenticationMiddleware(request);
-    const params = Object.fromEntries(request.nextUrl.searchParams.entries());
-    const pagination = getPaginationFromParams(params);
-
-    const invitations = await getTeamInvitations(params, pagination);
-
-    return Response.json(invitations);
-  } catch (error: unknown) {
-    console.error({
-      error,
-      step: '[My Team Invitations] Get team invitations by the logged user',
-    });
-    const message = isErrorWithMessage(error) ? error?.message : '';
-    return Response.json({ message }, { status: 400 });
-  }
-};
 
 export const POST = async (request: NextRequest) => {
   try {
