+++ src/controllers/teamCollaborator.controller.ts
@@ -1,11 +1,21 @@
 import { Attributes } from 'sequelize';
+import isEmail from 'validator/lib/isEmail';
 
 import { FilterObject } from '@/utils/filter';
-import { findTeamCollaboratorByUserId } from '@/services/teamCollaborator.service';
+import {
+  findTeamCollaboratorByUserId,
+  markAsAccepted,
+} from '@/services/teamCollaborator.service';
 import { PaginationOptions } from '@/utils/paginateData';
-import { TeamCollaborator, TeamRoles } from '@/models/teamCollaborator.model';
+import {
+  InvitationStatuses,
+  TeamCollaborator,
+  TeamRoles,
+} from '@/models/teamCollaborator.model';
 import { User } from '@/models/user.model';
 import { ValidatorError } from '@/utils/error.utils';
+import { validateMandatoryFields } from '@/utils/vaslidations';
+import { findUserByEmail } from '@/services/user.service';
 
 export async function getTeamCollaborators(
   filter: FilterObject,
@@ -30,12 +40,6 @@ export async function getMyTeamCollaborators(
 
 export async function findTeamCollaboratorById(id: string) {
   return TeamCollaborator.findOne({ where: { id } });
-}
-
-export async function addTeamCollaborator(
-  teamCollaboratorData: Attributes<TeamCollaborator>
-) {
-  return TeamCollaborator.create(teamCollaboratorData);
 }
 
 export async function updateTeamCollaboratorById(
@@ -80,4 +84,60 @@ export const removeMyCollaboratorById = async (
 
 export const findMyTeam = async (userId: string) => {
   return TeamCollaborator.findOne({ where: { user_id: userId } });
+};
+
+export const findTeamInvitationByEmail = (email: string) => {
+  return TeamCollaborator.findOne({ where: { email } });
+};
+
+export const addTeamCollaborator = async (
+  inputTeamCollaborator: Attributes<TeamCollaborator>
+) => {
+  const email = inputTeamCollaborator?.email;
+  const isValid =
+    validateMandatoryFields(inputTeamCollaborator, [
+      'team_id',
+      'email',
+      'status',
+    ]) && isEmail(email ?? '');
+
+  if (!isValid)
+    throw new ValidatorError('Team invitation params are not valid');
+
+  const hasInvitation = Boolean(await findTeamInvitationByEmail(email ?? ''));
+  const isEmailRegistered = Boolean(await findUserByEmail(email ?? ''));
+  if (hasInvitation || isEmailRegistered)
+    throw new ValidatorError('The email provided has an invitation already');
+
+  return TeamCollaborator.create(inputTeamCollaborator);
+};
+
+export const getCollaboratorTeam = async (id: string) => {
+  return TeamCollaborator.findOne({ where: { user_id: id } });
+};
+
+export const invitePersonToMyTeam = async (
+  user: User,
+  email: string,
+  role: string
+) => {
+  const expirationDate = new Date();
+  expirationDate.setDate(expirationDate.getDate() + 7);
+
+  const team = await getCollaboratorTeam(user?.id ?? '');
+  return await addTeamCollaborator({
+    team_id: team?.team_id ?? '',
+    email,
+    role,
+    status: InvitationStatuses.PENDING,
+  });
+};
+
+export const acceptTeamInvitation = async (user: User, isNew: boolean) => {
+  if (!isNew) return null;
+
+  const teamInvitation = await findTeamInvitationByEmail(user?.email);
+
+  if (teamInvitation?.status === InvitationStatuses.PENDING)
+    await markAsAccepted(teamInvitation?.id as string, user);
 };
