+++ src/services/teamCollaborator.service.ts
@@ -1,4 +1,8 @@
-import { TeamCollaborator } from '@/models/teamCollaborator.model';
+import {
+  InvitationStatuses,
+  TeamCollaborator,
+} from '@/models/teamCollaborator.model';
+import { User } from '@/models/user.model';
 import { Attributes } from 'sequelize';
 
 export const getTeamCollaborator = async (id: string) => {
@@ -17,4 +21,16 @@ export const addTeamCollaborator = async (
   teamCollaborator: Attributes<TeamCollaborator>
 ) => {
   return TeamCollaborator.create(teamCollaborator);
+};
+
+export const markAsAccepted = async (id: string, user: User) => {
+  return TeamCollaborator.update(
+    {
+      status: InvitationStatuses.ACCEPTED,
+      user_id: user?.id ?? '',
+    },
+    {
+      where: { id: id },
+    }
+  );
 };
